feat: archon hunt panel
