+++ static/lang/en.json
@@ -154,7 +154,10 @@
     "prediction": "Prediction"
   },
   "sortie": {
-    "header": "Sortie"
+    "header": "Sortie",
+    "hunt": {
+      "header": "Archon Hunt"
+    }
   },
   "steelPath": {
     "header": "Steel Path"
